+++ src/azure-cli/HISTORY.rst
@@ -15,6 +15,10 @@ Release History
 
 * az network private-dns link vnet create/update: Fixes #9851. Support cross-tenant virtual network linking.
 
+**Storage**
+
+* az storage copy: Add --s2sPreserveAccessTier parameter to preserve access tier during service to service copy.
+
 2.0.74
 ++++++